Refactorizacion final y últimas correcciones.
