+++ modulos/datos_dispositivos.py
